add if task destroy

+++ app/Http/Controllers/TasksController.php
@@ -7,11 +7,7 @@ use App\Task;
 
 class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         // $tasks = Task::all();
@@ -41,12 +37,6 @@ class TasksController extends Controller
             'task'=>$task]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $this->validate($request,[
@@ -63,39 +53,26 @@ class TasksController extends Controller
         return redirect('/');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function show($id)
     {
         $task = Task::find($id);
+        if(\Auth::id()=== $task->user_id){
         return view('tasks.show',[
             'task'=>$task]);
+        }
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         $task = Task::find($id);
+        if(\Auth::id()=== $task->user_id){
         return view('tasks.edit',[
             'task'=>$task]);
+        }
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+ 
     public function update(Request $request, $id)
     {   
         $this->validate($request,[
@@ -103,22 +80,21 @@ class TasksController extends Controller
             'status'=>'required|max:10']);
             
         $task= Task::find($id);
+         if(\Auth::id()=== $task->user_id){
         $task->content=$request->content;
         $task->status=$request->status;
         $task->save();
+         }
         return redirect('/');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
         $task = Task::find($id);
+        if(\Auth::id()=== $task->user_id){
         $task->delete();
+        }
         return redirect('/');
     }
 }
